moves BaselineCalcType enum
