+++ asf_search/Products/ALOSProduct.py
@@ -1,7 +1,7 @@
 import copy
 from asf_search import ASFSession, ASFProduct, ASFSearchOptions
 from asf_search.CMR.translate import get as umm_get, cast as umm_cast, try_parse_float, try_parse_int, try_round_float
-from asf_search.baseline import BaselineCalcType
+from asf_search.Products import BaselineCalcType
 from asf_search.exceptions import ASFBaselineError
 
 class ALOSProduct(ASFProduct):
