+++ asf_search/Products/ERSProduct.py
@@ -1,6 +1,6 @@
 import copy
 from asf_search import ASFSearchOptions, ASFSession, ASFProduct
-from asf_search.baseline import BaselineCalcType
+from asf_search.Products import BaselineCalcType
 from asf_search.CMR.translate import get as umm_get, cast as umm_cast, try_parse_float, try_round_float
 from asf_search.exceptions import ASFBaselineError
