+++ asf_search/baseline/__init__.py
@@ -1,3 +1,2 @@
 from .calc import *
 from .stack import *
-from .BaselineCalcType import BaselineCalcType
